refactor: move main bot class into core directory
