+++ kusogaki_bot/core/bot.py
